LP#1070466: Print to STDERR in Sip/Checksum::debug_split_print.
debug_split_print was printing to STDOUT, which goes to the client. This
causes the client to hang up because it is not receiving the expected
response from the server.  Explicitly printing to STDERR in this
function resolves this and you can continue your session while also
debugging checksum calculations.

Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>
Signed-off-by: Galen Charlton <gmc@esilibrary.com>