LP#1338761: add support for Relais extensions to message 24
The Relais SIP2 client expects that the Patron Status Response
(msg 24) will contain certain non-standard fields that are more
properly included in the patron information response:

- patron home address
- patron email
- patron home phone

In addition, and contrary to the specification, Relais expects that
the variable fields in the response will be given in the order that
they appear in the spec.

This patch ensures that the Patron Status repsonse emits the
institution ID, patron ID, and personal name in that order.  It also
adds a new institution option, relais_extensions_to_msg24, to make
the response includes the additional fields.  For example:

  <relais_extensions_to_msg24 enabled="true" />

Signed-off-by: Galen Charlton <gmc@esilibrary.com>