LP#1180468: update license statement to GPL v2 or later
In 2011, all current contributors and copyright holders
to SIPServer approved switching the license from GPLv2
to GPLv2 or later:

<http://libmail.georgialibraries.org/pipermail/open-ils-general/2011-April/004657.html>k

In 2013, contributor Jeff Godin also confirmed approval of
the license change:

https://bugs.launchpad.net/sipserver/+bug/1180468/comments/2

All subsequent contributors have either already approved of the
license switch or have done work-for-hire coding for organizations
that have approved the change.

Consequently, this patch adjusts all references to the GPLv2
to specify that the current license is GPLv2 or later.  This
patch also adds copyright and copying statements to a few
files that lacked them.

Signed-off-by: Galen Charlton <gmc@esilibrary.com>
Signed-off-by: Jeff Godin <jgodin@tadl.org>