LP 1463943: Improve Encoding Discovery
Improve the previous commit by looking for the encoding in the
institution implementation config in addition to the account
configuration.  The account configuration is used first, followed by
that from the institution implemenation config.

We also set a convenience field on the server object in the
_load_ils_handler method implementation.  This way, we only need get
the encoding once per session.  Hopefully this is correct with the
multiplex server.

Signed-off-by: Jason Stephenson <jason@sigio.com>
Signed-off-by: Martha Driscoll <driscoll@noblenet.org>
Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>