LP 1463943: More Encoding Improvements.
First, we improve the encoding search by looking in the institution
config proper.  The search order for an encoding is now:

1. An encoding attribute specified in the account tag.

2. An encoding tag in the institution tag.

3. An encoding tag in the implementation_config tag of the institution
tag as a last resort.

4. Use ASCII if none is set.  (The SIP2 specification practially
requires this.)

The last is done for backward compatibility with existing Evergreen
oils_sip.xml files.  This should be deprecated in the near future and
only the account tag and institution tag proper should be supported.

Second, we improve the handling of ASCII output by stealing a code
chunk from Evergreen's OILS::SIP->clean_text routine to remove
combining characters on strings.

Third, we remove the encoding attribute from the sample SIPconfig.xml
account tag for lpl-sc, and put an encoding tag in the institution
tags for LPL and UWOLS.

Signed-off-by: Jason Stephenson <jason@sigio.com>
Signed-off-by: Martha Driscoll <driscoll@noblenet.org>
Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>